Move InfoPanel to experimental
