+++ src/experimental/InfoPanel.tsx
@@ -2,20 +2,14 @@ import { Component } from 'solid-js';
 import './_infoPanel.css';
 import TextSample from '../composables/TextSample';
 
-const content = {
-    title: 'Share your Decentralized ID',
-    body: 'Copy your DID and use it to identify yourself anywhere within the Verifiable Credentials ecosystem.'
-}
-
 const InfoPanel: Component<{did: string}> = (props) => {
-
     return (
         <aside class="info-panel">
             <div>
                 <div>
                     <h2 class="overline">Quick Access</h2>
-                    <h3>{content.title}</h3>
-                    <p>{content.body}</p>
+                    <h3>Share your Decentralized ID</h3>
+                    <p>Copy your DID and use it to identify yourself anywhere within the Verifiable Credentials ecosystem.</p>
                 </div>
                 <div>
                     <div>
