+++ src/experimental/_infoPanel.css
